split out fused ops into their own files
